# docs/openapi.yaml
openapi: 3.0.3
info:
  title: OpenMuara
  description: Local-first billing and payment virtualization layer.
  version: 0.1.1
  contact:
    name: OpenMuara
servers:
  - url: http://127.0.0.1:9000
    description: Local development server
paths:
...
  /v1/pay:
    post:
      summary: Create a provider-agnostic payment
      operationId: createPayment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentRequest'
      responses:
        '201':
          description: Payment created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResponse'
        '200':
          description: Existing payment returned (idempotency hit)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /_admin/fawry-escape:
    get:
      summary: Fawry 3-D Secure escape simulation page
      operationId: fawryEscapePage
      parameters:
        - name: ref
          in: query
          required: true
          schema:
            type: string
        - name: returnUrl
          in: query
          required: true
          schema:
            type: string
        - name: amount
          in: query
          schema:
            type: string
      responses:
        '200':
          description: HTML escape page
...
  /v1/checkout/sessions:
    post:
      summary: Create a Stripe Checkout session
      operationId: stripeCreateCheckoutSession
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StripeCreateCheckoutSessionRequest'
      responses:
        '201':
          description: Session created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StripeCheckoutSession'
...
  /_admin/stripe/success:
    post:
      summary: Simulate Stripe checkout success
      operationId: stripeSimulateSuccess
      parameters:
        - name: session_id
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Session marked complete
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StripeCheckoutSession'
  /_admin/stripe/fail:
    post:
      summary: Simulate Stripe checkout failure
      operationId: stripeSimulateFail
      parameters:
        - name: session_id
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Session marked failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StripeCheckoutSession'
  /_admin/stripe/cancel:
    post:
      summary: Simulate Stripe checkout cancellation
      operationId: stripeSimulateCancel
      parameters:
        - name: session_id
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Session marked cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StripeCheckoutSession'
...
components:
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            trace_id:
              type: string
    TransactionItem:
      type: object
      properties:
        itemCode:
          type: string
        price:
          type: number
          format: double
        quantity:
          type: integer
    Transaction:
      type: object
      properties:
        id:
          type: string
        provider:
          type: string
        type:
          type: string
        amount:
          type: number
          format: double
        currency:
          type: string
        status:
          type: string
          enum: [new, paid, unpaid, refunded]
        customerRef:
          type: string
        idempotencyKey:
          type: string
        reference:
          type: string
        items:
          type: array
          items:
            $ref: '#/components/schemas/TransactionItem'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    PaymentRequest:
      type: object
      required: [provider, type, amount, currency, reference]
      properties:
        provider:
          type: string
        type:
          type: string
        amount:
          type: number
          format: double
        currency:
          type: string
        customer_ref:
          type: string
        reference:
          type: string
        idempotency_key:
          type: string
        items:
          type: array
          items:
            $ref: '#/components/schemas/TransactionItem'
    PaymentResponse:
      allOf:
        - $ref: '#/components/schemas/Transaction'
...
    StripeProductData:
      type: object
      properties:
        name:
          type: string
    StripePriceData:
      type: object
      properties:
        currency:
          type: string
        unit_amount:
          type: integer
        product_data:
          $ref: '#/components/schemas/StripeProductData'
    StripeLineItem:
      type: object
      properties:
        price_data:
          $ref: '#/components/schemas/StripePriceData'
        quantity:
          type: integer
    StripeCreateCheckoutSessionRequest:
      type: object
      required: [success_url, line_items, mode]
      properties:
        success_url:
          type: string
        cancel_url:
          type: string
        line_items:
          type: array
          items:
            $ref: '#/components/schemas/StripeLineItem'
        mode:
          type: string
        payment_method_types:
          type: array
          items:
            type: string
            enum: [card, fpx]
        customer_email:
          type: string
        client_reference_id:
          type: string
        metadata:
          type: object
          additionalProperties:
            type: string
    StripeCheckoutSession:
      type: object
      properties:
        id:
          type: string
        object:
          type: string
        amount_total:
          type: integer
        currency:
          type: string
        customer_email:
          type: string
        mode:
          type: string
        payment_method_types:
          type: array
          items:
            type: string
            enum: [card, fpx]
        payment_status:
          type: string
        status:
          type: string
        success_url:
          type: string
        cancel_url:
          type: string
        url:
          type: string
        client_reference_id:
          type: string
        metadata:
          type: object
          additionalProperties:
            type: string
...